Conflicts:
	bin/tests/AirCorridorInTimeTest.class
	bin/tests/ReadAirplanesCoordinatesTest.class
	bin/tests/ReportEmitterTest.class
	bin/utils/AirCorridorInTime.class
	bin/utils/Airship.class
	bin/utils/AltitudeComparator.class
	bin/utils/CargoAircraft.class
	bin/utils/CivilAirplane.class
	bin/utils/Database.class
	bin/utils/ReadAirplanesCoordinates.class
	bin/utils/ReadListOfFlights.class
	bin/utils/ReportEmitter.class
	src/utils/AltitudeComparator.java
	src/utils/CargoAircraft.java
	src/utils/CivilAirplane.java
	src/utils/ReadAirplanesCoordinates.java
	src/utils/ReadListOfFlights.java
	src/utils/ReportEmitter.java

+++ src/utils/Database.java
@@ -5,6 +5,9 @@ import java.util.Iterator;
 import java.util.Map;
 import java.util.Set;
 
+import app.InvalidArgumentException;
+import app.InvalidFlightIDException;
+
 /**
  * Allows to create a database of airships
  * 
@@ -31,9 +34,17 @@ public class Database {
 	 * @param airplane - airplane to add
 	 * @return true if the airplane was successfully added
 	 * @return false if the airplane was not added
+	 * @throws InvalidArgumentException 
 	 */
-	public boolean addAirplane(Airship airplane)
+	public boolean addAirplane(Airship airplane) throws InvalidFlightIDException, InvalidArgumentException
 	{
+		if(airplane == null)
+			throw new InvalidArgumentException();
+		
+		String id = airplane.getFlightID();
+		if (id == null)
+			throw new InvalidFlightIDException();
+		
 		if(database.containsKey(airplane.getFlightID()))
 		{
 			return false;
@@ -51,8 +62,11 @@ public class Database {
 	 * @return true if the airplane was successfully removed
 	 * @return false if the specified ID was not found in the database
 	 */
-	public boolean removeAirplane(String id)
+	public boolean removeAirplane(String id) throws InvalidFlightIDException
 	{
+		if(id ==null)
+			throw new InvalidFlightIDException();
+			
 		if(database.containsKey(id))
 		{
 			database.remove(id);
@@ -67,10 +81,17 @@ public class Database {
 	 * @param airplane - airplane to remove
 	 * @return true if the airplane was successfully removed
 	 * @return false if the airplane was not found in the database
+	 * @throws InvalidArgumentException 
 	 */
-	public boolean removeAirplane(Airship airplane)
+	public boolean removeAirplane(Airship airplane) throws InvalidFlightIDException, InvalidArgumentException
 	{
+		if(airplane == null)
+			throw new InvalidArgumentException();
+		
 		String id = airplane.getFlightID();
+			if(id == null) 
+				throw new InvalidFlightIDException();
+			
 		return removeAirplane(id);
 	}
 	
@@ -79,7 +100,7 @@ public class Database {
 	 * to have 0 passengers 
 	 * @return the number of airplanes that were removed from the database
 	 */
-	public int removeAirplanesWithZeroPassengers()
+	public int removeAirplanesWithZeroPassengers() throws InvalidFlightIDException
 	{
 		Set<String> idSet = database.keySet();
 		Iterator<String> iterator = idSet.iterator();
@@ -91,8 +112,12 @@ public class Database {
 			Airship airplane = database.get(iterator.next());
 			if((airplane instanceof utils.Airliner) && ((Airliner) airplane).getPassengersNumber() == 0)
 			{
+				String id = airplane.getFlightID();
+				if(id == null) 
+					throw new InvalidFlightIDException();
 				toRemove.add(airplane.getFlightID());
 				count++;
+				
 			}
 		}
 		
@@ -129,8 +154,10 @@ public class Database {
 	 * @param id the airplane id
 	 * @return true if the id is on database else returns false
 	 */
-	public boolean contains (String id)
+	public boolean contains (String id) throws InvalidFlightIDException
 	{
+		if(id == null) 
+			throw new InvalidFlightIDException();
 		return database.containsKey(id);
 	}
